requirements.txt
msal
pandas
python-dotenv
requests
